Make coord as doubles instead of int
In order to manage only coord as units, floats are needed to determine
sub-tile coord.

+++ coord.h
@@ -17,8 +17,8 @@ enum Cardinals {N, NE, E, SE, S, SW, W, NW};
 typedef struct Coord Coord;
 struct Coord
 {
-    unsigned int x;
-    unsigned int y;
+    double x;
+    double y;
     Bool pixels;
 };
